+++ packages/editor/web/src/RichContentEditor/RichContentEditor.jsx
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import { EditorState, convertFromRaw, Modifier } from 'draft-js';
 import Editor from 'draft-js-plugins-editor';
 import { get, includes, merge, debounce } from 'lodash';
 import Measure from 'react-measure';
@@ -15,12 +14,15 @@ import getBlockRenderMap from './getBlockRenderMap';
 import { combineStyleFns } from './combineStyleFns';
 import { getStaticTextToolbarId } from './Toolbars/toolbar-id';
 import {
+  EditorState,
+  convertFromRaw,
   TooltipHost,
   TOOLBARS,
   getBlockInfo,
   getFocusedBlockKey,
   calculateDiff,
   getPostContentSummary,
+  Modifier,
 } from 'wix-rich-content-editor-common';
 
 import {