+++ CHANGELOG.md
@@ -13,6 +13,10 @@
     Click to see more.
   </summary>
 
+</details>
+<hr/>
+
+## 7.0.0 (Mar 17, 2020)
 ### :boom: Breaking Change
 - `editor`
   - [#752](https://github.com/wix-incubator/rich-content/pull/752) Move draft-js to dependency from peerDependency [Migration Detials](https://github.com/wix-incubator/rich-content/wiki/RCE-V.7-Migration-Guide) 
@@ -24,9 +28,6 @@
 ### :bug: Bug Fix
 - `gallery`
   - [#775](https://github.com/wix-incubator/rich-content/pull/775) adding videos to gallery
-
-</details>
-<hr/>
 
 ## 6.10.2 (Mar 15, 2020)
 ### :house: Internal